backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
psycopg2-binary
django-cors-headers
django-filter
django-environ
django-axes
gunicorn
whitenoise[brotli]
Pillow
